requirements.txt
Django==1.4
MySQL-python==1.2.2
PIL==1.2a0
South==0.7.3
django-devserver==0.3.2
django-piston==0.2.3
django-registration==0.7
django-tagging==0.3.1
django-tastypie==0.9.11
requests==0.11.2
lxml==2.3.4
pystache==0.5.2
